Add JAX path for `convolve_image`
These changes allow the PSF convolution to be calculated in a JAX friendly way that works with `jax.grad`.  The method will take `image` and `blurring_image`, convert them from their 1D slim versions back into their 2D naive arrays (with masked points set to zero) and use `jax.scipy.signal.convolve` to do the convolution.

A new keyword has been added to the function `jax_method` that can be either `direct` or `fft` to control how the convolution is done in JAX.  Typically if the PSF kernel is more than about 5x5 the FFT method will be faster.  Because of this `fft` is the default.

Note: with the FFT some more speed might be gained by pre-caching the FFT of the PSF kernel as that is fixed and does not need to be re-computed every time, but this is an optimization that can be put in place at a later time once we know the current function works as expected.