Adjust numpy to jax.numpy imports for files used for the MGE Sersic
These files are all used by the Sersic profile's MGE deflection angles.  This wholesale replacement of numpy for jax.numpy solves many of the issues around the array wrappers (e.g. "Grid2d").  The `numpy_wrapper` is just a simple file that can be imported to provide the correct `np` and an optional `use_jax` bool in case further `if` blocks are needed.

Thankfully, with the full replacement much of the control flow `if` blocks seem to work without modification!